+++ src/app/app.toast.options.ts
@@ -1,0 +1,13 @@
+
+import { ToastOptions } from 'ng2-toastr';
+
+export class CustomOptions extends ToastOptions {
+  animate = 'fade';
+  dismiss = 'auto';
+  showCloseButton = true;
+  newestOnTop = true;
+  enableHTML = true;
+  toastLife = 5000;
+  // messageClass = '';
+  // titleClass = '';
+}
