+++ src/app/components/pin-details/gathering/participant-details/participant-details.component.ts
@@ -123,12 +123,12 @@ export class ParticipantDetailsComponent implements OnInit {
     }
 
     if (this.selectedRole === GroupRole.LEADER && this.leaderCount >= MaxGroupLeaders) {
-      this.toast.warning(this.content.getContent('finderMaxLeadersExceeded'), null, {toastLife: 3000});
+      this.toast.warning(this.content.getContent('finderMaxLeadersExceeded'), null);
       return;
     }
 
     if (this.selectedRole === GroupRole.APPRENTICE && this.apprenticeCount >= MaxGroupApprentices) {
-      this.toast.warning(this.content.getContent('finderMaxApprenticeExceeded'), null, {toastLife: 3000});
+      this.toast.warning(this.content.getContent('finderMaxApprenticeExceeded'), null);
       return;
     }
 
@@ -140,7 +140,7 @@ export class ParticipantDetailsComponent implements OnInit {
         this.router.navigate(['/small-group/' + this.groupId]);
       },
       failure => {
-        this.toast.warning('Something seems to have gone wrong. Please try again.', null, {toastLife: 3000});
+        this.toast.warning('Something seems to have gone wrong. Please try again.', null);
         this.state.setLoading(false);
       }
     );
